requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==4.2.17
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
oauthlib==3.2.2
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
